replace middlevare for localization
