+++ src/middleware_ignoring.ts
@@ -1,0 +1,49 @@
+import { getToken } from "next-auth/jwt";
+import { withAuth } from "next-auth/middleware";
+import { NextResponse } from "next/server";
+
+export default withAuth(
+  async function middleware(req) {
+    const pathname = req.nextUrl.pathname;
+
+    // Manage route protection
+    const isAuth = await getToken({ req });
+    const isLoginPage = pathname.startsWith("/login");
+
+    const sensitiveRoutes = ["/dashboard"];
+    const isAccessingSensitiveRoute = sensitiveRoutes.some((route) =>
+      pathname.startsWith(route)
+    );
+
+    if (isLoginPage) {
+      if (isAuth) {
+        return NextResponse.redirect(new URL("/dashboard", req.url));
+      }
+
+      return NextResponse.next();
+    }
+
+    if (!isAuth && isAccessingSensitiveRoute) {
+      return NextResponse.redirect(new URL("/login", req.url));
+    }
+
+    if (pathname === "/") {
+      return NextResponse.redirect(new URL("/dashboard", req.url));
+    }
+  },
+  {
+    // callbacks for call widleware every time
+    callbacks: {
+      async authorized() {
+        return true;
+      },
+    },
+  }
+);
+
+/**
+ * With what pages middleware will work
+ */
+export const config = {
+  matchter: ["/", "/login", "/dashboard/:path*"],
+};
